Move Source and other SQLA models out of experiment.py in all demos
